Refactor loop away from Dimensions
